requirements.txt
#object detection

bcrypt==4.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cvzone==1.5.6
cycler==0.12.1
deep-sort-realtime==1.3.2
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
intel-openmp==2021.4.0
Jinja2==3.1.6
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mkl==2021.4.0
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.5
numpy==1.26.4
opencv-python==4.7.0.72
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.16.0
six==1.17.0
sympy==1.14.0
tbb==2021.13.1
tifffile==2025.6.11
torch==2.3.1+cu118
torchaudio==2.7.1
torchvision==0.18.1+cu118
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0